feat: show the articles in the view and store extract in indexedDB
Change the api for the articles, getting them directly from Wikipedia because its supports https.  In the wikipedia api extract from the article needs to be obtained with a new api call(implemented). Subdivide the method to call the api and the one to update distances to have different polling times.